LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/sticky/view-timeline-sticky-offscreen-1-expected.txt

Subject

FAIL View timeline top-sticky during entry. assert_equals: Effect at the start of the active phase: cover 0% to cover 100% expected "0.3" but got "0.466667"
